pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "A static blog generator with markdown support"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "Apache-2.0"}
authors = [
    {name = "Roger Willis"}
]
dependencies = [
    "markdown>=3.5.2",
    "jinja2>=3.1.3",
    "python-slugify>=8.0.4",
    "markdown-captions>=2.1.2",
    "pygments>=2.17.2",
    "Pillow>=10.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["scripts"]

[tool.black]
line-length = 80
